+++ core/src/lib.rs
@@ -14,7 +14,6 @@ pub mod broadcast_stage;
 pub mod cache_block_meta_service;
 pub mod cluster_info_vote_listener;
 pub mod cluster_nodes;
-pub mod cluster_nodes_utils;
 pub mod cluster_slot_state_verifier;
 pub mod cluster_slots;
 pub mod cluster_slots_service;
