+++ core/src/retransmit_stage.rs
@@ -416,6 +416,7 @@ pub fn retransmitter(
         .unwrap()
 }
 
+// pub for bench
 pub struct RetransmitStage {
     retransmit_thread_handle: JoinHandle<()>,
     window_service: WindowService,
